### 2017-09-05 #393 @whedon
[PRE REVIEW]: bittrex: A R Client for the Bittrex Crypto-Currency Exchange
**Submitting author:** @kaneplusplus (<a href="http://orcid.org/0000-0003-1899-6662">Michael Kane</a>)
**Repository:** <a href="https://github.com/ropensci/bittrex/" target ="_blank">https://github.com/ropensci/bittrex/</a>
**Version:** v0.1.6
