### 2017-09-05 #393 @arfon
👋 @kaneplusplus - thanks for your submission to JOSS. We've recently tightened up our submission criteria for JOSS (this pre-dates your submission to rOpensci). Would you be able to add some extra language to this `paper.md` to address the [submission criteria](http://joss.theoj.org/about#submission_requirements). In particular, this item:

> - Your software should have an obvious research application

Could you confirm here that there _is_ a research application for this software (and explain what that application is in the paper)? The section [_'what should my paper contain'_](http://joss.theoj.org/about#paper_structure) has some guidance for the sort of content we're looking to be present in the `paper.md`.

Many thanks!
