### 2017-09-06 #393 @kaneplusplus
@arfon Yes, this is not problem and I'm happy to make this clearer. I confirm this does have research applications. A few financial/economics research areas and topics include:
- Comparisons between semi-regulated markets (bittrex) to traditional, regulated markets.
- Microfinancing for start-ups using initial coin offerings.
- Quantifying the value of non-traditional products and services.
- Systemic risk of multiple crypto-currencies on the same block chain.
- Crypto-coins as a proxy for sports betting where sports betting is illegal.

I will clean this up and revise `paper.md` in the next couple of days and follow up.
