### 2017-09-11 #393 @arfon
@kaneplusplus - could you link to the rOpensci onboarding thread here please?
