### 2017-09-11 #393 @karthik
@arfon https://github.com/ropensci/onboarding/issues/120
