### 2017-09-12 #402 @arfon
Accepting without further review because of submission to rOpensci https://github.com/ropensci/onboarding/issues/120
