### 2017-09-12 #402 @arfon
@kaneplusplus - please merge this PR https://github.com/ropensci/bittrex/pull/6 and then make an archive of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive. I can then move forward with accepting the submission.
