### 2017-09-12 #402 @kaneplusplus
@arfon I've made the merge and updated in Zenodo. The DOI is 10.5281 (zenodo.889710). Thanks very much.
