### 2017-09-13 #402 @arfon
@kaneplusplus - your paper is now accepted into JOSS and your DOI is http://dx.doi.org/10.21105/joss.00402 :zap: 🚀 💥
